updated entities

+++ src/Domain/Entities/QuestionEntity.cs
@@ -4,6 +4,6 @@ public class QuestionEntity : BaseEntity
 {
     public string? Text { get; set; }
     public TestEntity? Test { get; set; }
-    public double? Mark { get; set; }
+    public decimal? Mark { get; set; }
     public ICollection<AnswerEntity>? Answers { get; set; }
 }
